credit-memo-gen/server/requirements.txt
flask
flask-cors
pdfplumber
google-generativeai
python-dotenv